OneDrive/Desktop/IPD project/certificate_generator/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
mediapipe==0.10.8
transformers==4.36.2
torch==2.1.2
torchaudio==2.1.2
numpy==1.24.3
pandas==2.1.3
python-jose==3.3.0
python-multipart==0.0.6
pydantic==2.5.2
gunicorn==21.2.0
opencv-python==4.8.1.78
scikit-learn==1.3.2
librosa==0.10.1
soundfile==0.12.1
Pillow==10.2.0
Flask-WTF==1.2.1
email-validator==2.1.0.post1
pdfkit==1.0.0 